requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
psycopg2==2.9.1
pytz==2020.1
sqlparse==0.3.1
